Implemented Enemy actions and changed enemies classes

+++ src/enemies/Enemy.java
@@ -35,7 +35,7 @@ public interface Enemy {
      *
      * @return damage : Integer
      */
-    Integer attack();
+    Float attack();
 
     /**
      * Checks if a Enemy blocks
